Initial implementation of course instance exercise point list (#439)
* Initial implementation of course instance exercise point list

* Match point-list with design, improve backend

* Updated ts-rs, fixed point-list issues

* fixed accessibility issues

* Added system test

* Added numbers util and test

* Address feedback

* Fixed test

* Fixed system tests

+++ services/headless-lms/src/controllers/main_frontend/exercises.rs
@@ -15,7 +15,7 @@ use uuid::Uuid;
 #[derive(Debug, Serialize, TS)]
 pub struct ExerciseSubmissions {
     data: Vec<Submission>,
-    total_pages: i64,
+    total_pages: u32,
 }
 
 /**
